infra/codebuild/lambda/requirements.txt
requests==2.21.0
PyGithub==1.43.4
